criação e edição de ações
tive que comentar quase todos setModal(true), tava zoando a navegação,
vamos ter que repensar issae :/

+++ TCC/src/Modelo/Variavel.java
@@ -14,7 +14,7 @@ import java.io.Serializable;
 public class Variavel implements Serializable{
     
     private String nome;
-    private long valorInicial;
+    private double valorInicial;
     private boolean oculta;
 
     public String getNome() {
@@ -25,11 +25,11 @@ public class Variavel implements Serializable{
         this.nome = nome;
     }
 
-    public long getValorInicial() {
+    public double getValorInicial() {
         return valorInicial;
     }
 
-    public void setValorInicial(long valorInicial) {
+    public void setValorInicial(double valorInicial) {
         this.valorInicial = valorInicial;
     }
